Refactors Visualiser to support checking mechanisms for editor and selection without calculating the actual diagram text.
This change is introduced to conform to the PlantUML API. It may as well reduce computation time, although negligible. However,
it is important for understanding and maintaining the code to have these methods implemented as intended. Both methods clearly
describe which editors (or in this case their input) and which selections are supported.
supportsEditor(...) checks for proper file extensions of the file currently edited, e.g. "ecore", and as well whether or not
the editor internally handles EObjects via its EditingDomain. supportsSelection(...) checks the selection for elements that
cannot be displayed. If there are any, the selection will be reported as not supported.

Utility methods have been added with VisualiserUtilities. The intention of this class is to provide most of the behavior which
interacts with the Eclipse Platform API, to keep the Visualiser code as clean as possible. This should in turn improve the
maintainability of the Visualiser code.