Fixes bug where metamodel and model elements are visualised in the same diagram.
Reason for this bug is, that even if a selection is reported to be not supported (in contrast to the editor), PlantUML will
call getDiagramText(...) with a null selection. Semi-correctly, it was assumed that the editor output can be used to generate
the diagram text. However, it was not checked wether or not the editor output can be clearly classified to contain only
metamodel or model elements. After adding this check and performing some extra improvements on the EMoflonEcoreVisualiser code
the bug had been fixed.