Fixes bug in metamodel visualiser when selectioning only EPackages, EOperations, EParameters, EGenericTypes or EStructuralFeatures.
Previously, if one of the above element types has been selected in the metamodel visualiser, instead of its containing class
element, the EObject itself was visualised by the model visualiser or not at all, if it was contained in a selection consisting
of EObjects and EModelElements / EGenericTypes. This was the case because no representation for these elements were chosen,
only EClasses in a selection could be visualised by the metamodel visualiser. Now, EStructuralFeatures, EOperations and
EParameters are represented by their according containing EClasses. EPackages are represented by their contained EClassifiers
and those contained in their subpackages. An EGernericType is represented by its EClassifier.

One problem still remains, EDataTypes (a specialization of EClassifier) cannot be represented by an EClass. Which means, they
cannot be visualised, if selected (an empty PlantUML diagram will be returned).