Refactoring: Moves control over which element to draw to visualisers.
Before this commit, the PlantUML generator had partial control over which elements to visualise and which not. For example, the
superclass of an EClass was always visualised, including the generalisation dependency. This behavior may change in the future,
amongst other. The new agenda for the PlantUML generator is to preciselly draw what is given to it. All calculations for
determining superclasses have been added to the metamodel visualiser and removed from the generator. The generalisation
dependency is now given in the form of a VisualEdge. To differentiate between the different kinds of edges, an Enum has been
added, which for now differentiates between LINK (model), REFERENCE (metamodel) and GENERALISATION (metamodel). Accordingly,
the generation of edges in the PlantUML generator has been unified into one method.