Refactoring: Impelements strategy pattern for diagram creation in metamodel and model visualisers.
This refactoring includes various strategies for the computation of object and class diagrams by the model and meta model
visualiser. To simplify the chaining of multiple strategies for diagram manipulation, new container types are introduced:
Diagram, ClassDiagram and ObjectDiagram. Amongst others, they store the selection and neighbourhood of nodes, as well as the
edges between them. The PlantUML generator was adjusted to draw ClassDiagrams and ObjectDiagrams. Hence, future functional
enhancements might benefit from these new container classes. The new container types are important for strategy operations,
since chaining of strategies is more easily possible. Initial strategies include: edge creation for a selection of nodes and
bidirectional expansion of the current neighbourhood by a degree of one. If a higher degree is wished, the latter method can
simply be chained. The mentioned strategies have been added for object and class diagrams.

In proceeding of the refactoring, all List type-references have been changed to the more general Collection, whereas
internally HashMaps are used if a collection type is to be instantiated. This should improve the performance, as the order of
elements appears to be not of relevance.