Improvement: Makes eMoflon visualisers configurable.
If an eMoflon visualiser implements the ConfigurableVisualiser interface, it can be configured by the singleton Configurator
class. E.g. handlers can then instruct the configurator to set a diagram processing strategy or style bits for the PlantUML
code generation. Everything method that conforms to the functional interface DiagramStrategy may be passed to the configurator
as a strategy.

A DiagramStrategy takes a diagram as input, and returns the processed diagram. The actual type, or what
passes as a diagram, is not enforced. However, the ConfigurableVisualiser interface defines a generic type paramater to
associate a configurable visualiser with a diagram type and hence, the diagram type supported by a strategy for that
visualiser. For example, the metamodel visualiser uses ClassDiagram as container type for a class diagram. Therefore it will
only accept methods as strategy which take a ClassDiagram instance as input and return value.

The EMoflonPlantUMLGenerator has been adjusted to take style bits as an input for its metamodel and model generation methods.
However, no functionality has been implemented with it, as this will follow in later commits.