Adds "Show model details" button to view menu of PlantUML view.
This toggle button allows to switch between additional details for models or none. Internally, a style bit in combination with
the Configurator class has been used. The style bits are handed to the PlantUMLGenerator, where they are evaluated. In
addition, the PlantUMLGenerator has been extended to generate operations and attributes of EClass elements as well.