Adds ViewAction to PlantUML view to toggle abbreviation of labels in models.
For this addition, a new command, handler and menu contribution has been added to plugin.xml. The according handler class sets
a style bit in the Configurator instance. This stylebit is handed to the EMoflonPlantUMLGenerator via the configurable
metamodel and model visualisers. Internally the generator uses apache StringUtils to crop the middle of an oversized string. By
default, a string counts as oversized if its length exceeds 11 characters. This length bound is not configurable, however,
a realization using the preference store should be feasible.

In addition, the notation in model visualisations has been adjusted. Objects are now named <instance_name> : <class_name>
instead of <instance_name>.<class_name>.