Adds ViewAction for showing 1-neighbourhood for metmodel and model visualiser.
If the action is active, the associated handler sets the a the according bidirectional 1-neighbourhood strategy via the
configurator. The metamodel and model visualisers will execute this strategy, and the 1-neighbourhood for the current
selection will be added to the diagram. So far, this is only supported with the model and metamodel visualisers.
This commit does not include an enableWhen clause for the handler.
If one wishes to add an own handler, to enable the support for other visualisers as well,
this has to be done, e.g. via a property tester, which calls supportsEditor and supportsSelection on the visualisers. Only one
handler can be active for a command at the same time, which is why the included handler cannot be active in this case.