Adds "Show documentation" ViewAction to PlantUML view.
The metamodel visualiser was extended to support the extraction of documentation values from EAnnotations. A handler for a
ViewAction has been added, to toggle the visualisation of these documentation values in the PlantUML view. To toggle this, the
Handler instructs the Configurator to set a specific style bit.

In addition, a bug in the VisualiserUtilities has been fixed. Previously, EMap Entries where contained in an expanded Resource
/ ResourceSet, which couldn't be handled by the metamodel visualiser. This is has been fixed by removing these elements from
the result.