Changes colors for metamodel and model and reworks naming mechanism for objects in models.
The calculation for object naming is now performed in the model visualiser itself, the result is stored in the ObjectDiagram.
This is handed to the generator. The names are named like this <family1 : Family> instead of <o1 : Family>.
Object instance names are still not ordered according to their containment relation, this is future work. Instead, the numbers
trailing the Object instance names are unique, but randomly assigned to the object.

The colors for both metamodel and model have been adjusted to be black and white, with the exception of EPackage
representations in metamodels. They are colored in a light gray.