Fixes nondeterministic behavior for metamodel and model visualiser if both .xmi and .ecore Resources are selected.
The problem: If both .xmi and .ecore Resources were selected in a supported editor, either the EcoreVisualiser correctly
reported an error diagram or the plantuml.ecore plugin visualised one of the resources.

The cause: Both PlantUML ecore visualisers and both the emMoflon metamodel and model visualiser had the same "default"
priority. Dependending on the order, in which eMoflon and PlantUML ecore visualiser were loaded by the PlantUML.eclipse plugin,
either one of those was asked first to visualise. The order in which the PlantUML.eclipse Activator loads the visualisers
appears to be undeterministic. Furthermore, the supportsSelection check of the PlatnUML.ecore visualiser checks only the first
element of a selection. Which again is influenced by which Resource is selected first in the Editor. The process of choosing a
visualiser for getting the diagram text is devided into two phases: first supportsEditor and supportsSelection is called. If
supportsEditor returns true, but not supportsSelection, the second phase starts. There again supportsEditor and
supportsSelection are called, the latter with a null argument for the selection. Both PlantUML.ecore and eModflon visualisers
seem to return the same result in this case, i.e. true, such that the first visualiser in line is chosen for diagram text
generation.

The fix: The priority value of the eMoflon metamodel and model visualisers has been increased to "custom" (=10), in contrast to
"default" (=5). This way the eMoflon metamodel and model visualisers always chosen first for diagram text generation. Thus, the
behavior is deterministic.